loss curves and accuracy curves
to show that model is performing well with each successive epoch and the loss (especially the test loss) is going down. Moreover, the Accuracy of the model (Test accuracy to be specific) is going upwards Logarithmically). Also, the test loss is below the train loss which is ideal condition for a classification model.